requirements.txt
beautifulsoup4==4.11.1
certifi==2022.6.15
coreapi==2.3.3
coreschema==0.0.4
Django==4.0.6
django-bootstrap4==22.1
django-bootstrap-datepicker-plus==4.0.0
django-cors-headers==3.13.0
django-filebrowser-no-grappelli==4.0.1
django-filter==22.1
django-select2==7.10.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-spectacular==0.22.1
ExifRead==3.0.0
geographiclib==2.0
geopy==2.2.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
Pillow==9.2.0
psycopg2-binary==2.9.3
pyparsing==3.0.9
pytz==2022.1
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.10
